application/requirements.txt
azure_storage==0.37.0
fastapi==0.94.1
imagekitio==3.0.1
msrest==0.7.1
requests==2.26.0
SQLAlchemy==2.0.4
starlette==0.26.1
urllib3==1.26.7